refactor: removed interface GameObject. Renamed GameObjectImpl to GameObject
